The tf::TransformBroadcaster is unused and cause ros::init problems
There are currently no use for the tf::TransformBroadcaster in the plugin. Plus since it is an object, it get initialized (with its ros::NodeHandle) before the call to the class creator, which creates an undesired sequence of event with ros::init, that has to be called before creating any ros::NodeHandle.